Changes to SetupController.java and modifications to TowerEatsSetup.fxml
Removed imports that aren't needed.
Updated selectedTower from ArrayList to list.
Updated variable that stores players name to be more clear.
Added Lists that store difficultyButtons and difficulties.
Updated colour that button is to change when pressed for Pick Starting Towers buttons.
Added functionality to add selected towers to selectedTower and update Button text with ResourceType of the selected tower //TODO Player can't add multiple of the same tower.
RoundsSlider now stores the intValue of the slider to maxRoundNum.
 playerName will store what the player enters as their from NameTextField.
 difficulty will store Difficulty Easy, Medium, and Hard depending on what button the player presses.
 TowerEatsSetup.fxml - minor changes to scale of labels, Adjustments to how RoundSlider is displayed, and removed text in TowerDescription.