SetupController.java, ShopCellFactory.java, and Tower.java
SetupController.java - updated buttons for Pick Starting Towers to show the towers name and change selected Towers buttons from showing resource type to towers name.
ShopCellFactory.java - added an else if statement to check if instance is a tower or upgrade (WIP) and cast Tower to be able to access methods.
Tower.java - added a getter that returns the name of the tower.