MainController.java changes
Added functionality to be able to buy items from the shop (WIP) need to remove items from shop to inventory/reserve towers.
Added method that will display a text popup if you purchased or can't afford something from the shop for a couple of seconds.
(WIP) popup for round summary.